+++ pages/api/games.js
@@ -5,7 +5,7 @@ export default async function handler(req, res) {
   try {
     await dbConnect();
     const games = await Game.find();
-    res.json(games[1]);
+    res.json(games);
     // res.json(games);
     // res.send("hi");
   } catch (err) {
