+++ pages/games/[_id].js
